requirements.txt
Flask
qrcode
gunicorn
Pillow
opencv-python-headless
